requirements.txt
Flask
requests==2.27.1
pytest-flake8
Werkzeug==0.16.1
flake8
pytest-flask
gunicorn